Add ThemeContext
This is meant to be able to define a new default theme for all
destamatic-ui components. It also streamlines the implementation of
all new components so they don't have to worry about the theme themselves.

This allows quickly changing the theme of a set of components without defining
a whole new theme.